Move kubehelpers package to top-level
All packages in the project are not intended for public use. They do not
ensure backwards compatibility. For this reasons, this commit removes
the internal/ directory. Everything is internal.
